backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.1
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
Pillow==10.2.0
django-filter==23.5
drf-spectacular==0.27.1
phonenumbers==8.13.27
django-phonenumber-field==7.3.0